add shared
