+++ crates/cli/src/cli/commands.rs
@@ -1,15 +1,14 @@
-use std::path::PathBuf;
 use std::env;
 
 use vencord_installer_core::Error;
-use vencord_installer_core::paths::{branch::DiscordLocation, locations::get_data_path, shared::get_custom_discord_location};
+use vencord_installer_core::paths::{branch::DiscordLocation, shared::get_custom_discord_location};
 use vencord_installer_core::patch::{patch_mod::Installer, patch_openasar::OpenAsarInstaller};
-use vencord_installer_core::update::{download::prepare_dist_directory, version_check::{check_hash_from_release, check_local_version}};
+
+use vencord_installer_shared::{OPENASAR_URL, USER_AGENT};
+use vencord_installer_shared::{info, success, warn};
+use vencord_installer_shared::{download_assets, get_dist_path};
 
 use tokio::runtime::Runtime;
-
-use crate::{info, success, warn};
-use crate::{OPENASAR_URL, RELEASE_TAG_DOWNLOAD, RELEASE_URL, RELEASE_URL_FALLBACK, USER_AGENT};
 
 use super::selections::select_location;
 
@@ -114,48 +113,4 @@ pub fn uninstall(client_mod: bool, openasar: bool, custom_path: Option<String>) 
     }
 
     Ok(())
-}
-
-// MARK: - Download
-
-fn download_assets() -> Result<(), Error> {
-    let rt = Runtime::new().unwrap();
-
-    info!("Checking for dist files to download...");
-
-    let latest_version = rt.block_on(check_hash_from_release(RELEASE_URL, Some(RELEASE_URL_FALLBACK), USER_AGENT));
-    let local_version = rt.block_on(check_local_version(&get_dist_path(), r"// Vencord ([0-9a-zA-Z\.-]+)"));
-
-    info!("Latest version: {}", latest_version.clone().unwrap_or_default());
-    info!("Local version: {}", local_version.clone().unwrap_or_default());
-
-    if latest_version.is_some() && latest_version != local_version {
-        info!("Downloading dist files...");
-
-        rt.block_on(prepare_dist_directory(
-            &get_dist_path(),
-            RELEASE_TAG_DOWNLOAD,
-            USER_AGENT,
-            [
-                "patcher.js".to_string(),
-                "preload.js".to_string(),
-                "renderer.js".to_string(),
-                "renderer.css".to_string(),
-            ],
-        ))?;
-    } else {
-        info!("Nothing new to download, skipping!");
-    }
-
-    Ok(())
-}
-
-// MARK: - Paths
-
-fn get_dist_path() -> PathBuf {
-    if let Ok(path) = env::var("VENCORD_USER_DATA_DIR") {
-        PathBuf::from(path).join("dist")
-    } else {
-        get_data_path("Vencord")
-    }
 }
